requirements.txt
# Core AI & Orchestration
langgraph
langchain-google-genai
qdrant-client

# Data Engineering
unstructured[all-docs]
requests
pandas
python-dotenv
networkx

# Evaluation & API
ragas
fastapi
uvicorn
streamlit